requirements.txt
tkinter
pymysql
pycrypto
